fix(ui): open in IDE error styles
